Blind Signature Verification on the Smart Contract now works.
Implemented pkcs1v1.5 encoding of messages for this to work.  The
message to be signed is first padded by the pseudonym as per the
specification: https://tools.ietf.org/html/rfc3447#section-9.2

This encoded message is then blinded by the Psuedonym and sent to the
Patron for signing. The patron uses vanilla rsa exponentiation to
blindly sign this message. This signature is un-blinded by the
Psuedonym. The result of the un-unblinding is a valid pkcs1v1.5 RSA
Signature.

Smart Contract is now successfully verifies the pkcs1v1.5 Signature.